add rq module

+++ person/novelScratch/util.py
@@ -3,5 +3,5 @@
 # @param {seq} list
 # @param {num} equal nums
 def chunkIt(seq, num):
-  num = len(seq) / num
+  num = int(len(seq) / num)
   return [seq[i:i+num] for i in range(0, len(seq), num)]
